lambda/log-retention/requirements.txt
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6 